feat: 과목 조회 페이지에 API 구조 리팩토링 및 React Query 적용(#128)
- API 요청 구조 분리
- 과목 조회 시 API 요청 함수 및 React Query를 사용한 useSubjects.ts 작성
- App.tsx에 QueryClientProvider 설정 적용
- Quizzes의 import 부분 충돌 해결
