+++ src/App.tsx
@@ -10,11 +10,14 @@ import Dashboard from '@pages/quizzes/Dashboard'
 import Deployments from '@pages/quizzes/Deployments'
 import Quizzes from '@pages/quizzes/Quizzes'
 import Submissions from '@pages/quizzes/Submissions'
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
+import { useEffect } from 'react'
 import { Toaster } from 'react-hot-toast'
 import { Route, Routes, useNavigate } from 'react-router'
-import { setGlobalNavigator } from './utils/navigation'
-import { useEffect } from 'react'
 import './App.css'
+import { setGlobalNavigator } from './utils/navigation'
+
+const queryClient = new QueryClient()
 
 function App() {
   const navigate = useNavigate()
@@ -41,17 +44,19 @@ function App() {
 
   return (
     <>
-      <Routes>
-        <Route element={<Login />} path="/" />
+      <QueryClientProvider client={queryClient}>
+        <Routes>
+          <Route element={<Login />} path="/" />
 
-        <Route element={<Layout />}>
-          {ROUTES.map((route) => (
-            <Route key={route.path} {...route} />
-          ))}
-        </Route>
-      </Routes>
+          <Route element={<Layout />}>
+            {ROUTES.map((route) => (
+              <Route key={route.path} {...route} />
+            ))}
+          </Route>
+        </Routes>
 
-      <Toaster position="top-right" />
+        <Toaster position="top-right" />
+      </QueryClientProvider>{' '}
     </>
   )
 }
