+++ src/hooks/queries/useSubjects.ts
@@ -1,0 +1,9 @@
+import { fetchSubjects } from '@api/subjects/subjectAPI'
+import { useQuery } from '@tanstack/react-query'
+
+export const useSubjects = () => {
+  return useQuery({
+    queryKey: ['subjects'],
+    queryFn: fetchSubjects,
+  })
+}
