+++ src/pages/courses/Courses.tsx
@@ -1,17 +1,17 @@
+import api from '@api/instance/axiosInstance'
 import Button from '@components/common/Button'
 import DataTable from '@components/common/data-table/DataTable'
 import Pagination from '@components/common/data-table/Pagination'
+import FormRow from '@components/common/FormRow'
+import ImageUploader from '@components/common/ImageUploader'
+import Input from '@components/common/Input'
+import Modal from '@components/common/Modal'
 import { ADMIN_API_PATH } from '@constants/urls'
 import type { TableRowData } from '@custom-types/table'
 import { useClientPagination } from '@hooks/data-table/usePagination'
-import api from '@api/axiosInstance'
-import { useEffect, useState } from 'react'
-import FormRow from '@components/common/FormRow'
-import Input from '@components/common/Input'
-import Modal from '@components/common/Modal'
 import { useCustomToast } from '@hooks/toast/useToast'
 import { cn } from '@utils/cn'
-import ImageUploader from '@components/common/ImageUploader'
+import { useEffect, useState } from 'react'
 
 // 페이지 상수 추가
 const COUNT_LIMIT = 20
