+++ src/pages/quizzes/Deployments.tsx
@@ -1,22 +1,22 @@
+import api from '@api/instance/axiosInstance'
+import SearchIcon from '@assets/icons/search.svg?react'
 import Button from '@components/common/Button'
 import DataTable from '@components/common/data-table/DataTable'
-import Modal from '@components/common/Modal'
+import Pagination from '@components/common/data-table/Pagination'
 import DropdownField from '@components/common/DropdownField'
 import Icon from '@components/common/Icon'
-import SearchIcon from '@assets/icons/search.svg?react'
-import { useState, useMemo, useEffect } from 'react'
-import Pagination from '@components/common/data-table/Pagination'
-import { useSort } from '@hooks/data-table/useSort'
-import { useClientPagination } from '@hooks/data-table/usePagination'
+import Modal from '@components/common/Modal'
+import SearchBar from '@components/common/SearchBar'
+import { ADMIN_API_PATH } from '@constants/urls'
 import {
   mapDeployment,
   type Deployment,
   type DeploymentResponse,
 } from '@custom-types/deployments'
-import SearchBar from '@components/common/SearchBar'
-import api from '@api/axiosInstance'
+import { useClientPagination } from '@hooks/data-table/usePagination'
+import { useSort } from '@hooks/data-table/useSort'
 import { filterOption } from '@utils/filterOption'
-import { ADMIN_API_PATH } from '@constants/urls'
+import { useEffect, useMemo, useState } from 'react'
 
 // 페이지 상수
 const COUNT_LIMIT = 20
