+++ src/pages/quizzes/Quizzes.tsx
@@ -1,22 +1,22 @@
+import api from '@api/instance/axiosInstance'
 import FilterIcon from '@assets/icons/search.svg?react'
 import Button from '@components/common/Button'
 import DataTable from '@components/common/data-table/DataTable'
 import Pagination from '@components/common/data-table/Pagination'
 import Icon from '@components/common/Icon'
 import SearchBar from '@components/common/SearchBar'
-import type { TableRowData } from '@custom-types/table'
-import { useServerPagination } from '@hooks/data-table/usePagination'
-import { useSort } from '@hooks/data-table/useSort'
-import { useCallback, useEffect, useMemo, useState } from 'react'
-import { ADMIN_API_PATH } from '@constants/urls'
 import ScheduleModal from '@components/create-schedule/ScheduleModal'
-import { useScheduleStore } from '@store/create-schedule/scheduleStore'
-import { quizAPI } from '@lib/api/scheduleApi'
-import type { SchedulePayload } from '@custom-types/createSchedule'
-import api from '@api/axiosInstance'
 import AddExamModal from '@components/quizzes/add-exam-modal/AddExamModal'
 import CourseFilterModal from '@components/quizzes/course-filter-modal/CourseFilterModal'
 import DetailModal from '@components/quizzes/detail-modal/DetailModal'
+import { ADMIN_API_PATH } from '@constants/urls'
+import type { SchedulePayload } from '@custom-types/createSchedule'
+import type { TableRowData } from '@custom-types/table'
+import { useServerPagination } from '@hooks/data-table/usePagination'
+import { useSort } from '@hooks/data-table/useSort'
+import { quizAPI } from '@lib/api/scheduleApi'
+import { useScheduleStore } from '@store/create-schedule/scheduleStore'
+import { useCallback, useEffect, useMemo, useState } from 'react'
 
 // 표제목 상수화
 const TableHeaderItem = [
